+++ clients/admin-ui/src/types/api/models/MinimalPrivacyExperience.ts
@@ -1,0 +1,8 @@
+/* istanbul ignore file */
+/* tslint:disable */
+/* eslint-disable */
+
+export type MinimalPrivacyExperience = {
+  id: string;
+  name: string;
+};